+++ api/sixtyfps-rs/docs.rs
@@ -109,10 +109,10 @@ pub mod generated_code {
         }
         /// Assigns a new value to the `user_name` property.
         pub fn set_user_name(&self, value: re_exports::SharedString) {}
-        /// For each callback declared at the root of the component, a function to emit that
-        /// callback is generated. This is the function that emits the `hello` callback declared
+        /// For each callback declared at the root of the component, a function to call that
+        /// callback is generated. This is the function that calls the `hello` callback declared
         /// in the `.60` design markup.
-        pub fn emit_hello(&self) {}
+        pub fn call_hello(&self) {}
         /// For each callback declared at the root of the component, a function connect to that callback
         /// is generated. This is the function that registers the function f as callback when the
         /// callback `hello` is emitted. In order to access